Changed Text to Text_Basic and added Text_Plus, which allows you to use - text: One character (fixed width) - text: Normal text (can be monospaced) - line: A way to put multiple things back to back - multiline: A way to put multiple things above/below each other - size: A way to make things bigger or smaller (default is 1.0) - fraction: A way to put one thing above another, with a line separating the two right in the middle.
A small paragraph like this

- A headline double the size of the normal text, followed by (Author) at half size
- The body

Would be written like so:

multiline:
  line:
    size:2.0*text:Headline\:
    size:0.5*text:(Author)\:
  /
  text:Body\:
/

In one line:

multiline:line:size:2.0*text:Headline\:size:0.5*text:(Author)\:/text:Body\:/

Notes:

in text: \: Terminates the string, indicating the end.
in size: * separates the scale factor from the object to be scaled
in line and multiline: / terminates the line or multiline block. This is not necessary at the end, so line:line:// and line:line both work.
"line:" is necessary for the headline because there are two parts there, the bigger one (size:2.0*...) and the smaller one (size:0.5*...). For the body, this is not necessary as there is only one text block there.

Right click -> New -> Text_Plus will insert a text that already has some information prefilled.

There will be a GUI to edit text in the future, but it might take a while as I might not have the time or determination to write GUI code right now.